refactor!: use layout groups
W.I.P

Before you flip due to the large amount of changes: check the documentation for svelte-kit's new file based router. It forces every route to be inside a folder / `+page.svelte`. Named layouts have also been replaced with layout groups. That is why there are now two folders called `(reset)` and `(rgb)` directly inside the routes folder.
