+++ src/routes/(reset)/projects/+layout.svelte
