+++ src/routes/(reset)/projects/+page.svelte
