+++ src/routes/(reset)/projects/barcode/+page.svelte
