+++ src/routes/(reset)/projects/color-picker/+page.svelte
