+++ src/routes/(reset)/projects/dnd-dice/character-roller/+page.svelte
@@ -1,4 +1,6 @@
 <script>
+	throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
 	import { onMount } from 'svelte';
 
 	import Alert from '$lib/components/Alert.svelte';
