+++ src/routes/(reset)/projects/doodle-pad/+layout@(reset).svelte
