+++ src/routes/(reset)/projects/doodle-pad/+page.svelte
