+++ src/routes/(reset)/projects/simple-grade-calc/+page.svelte
