+++ src/routes/(rgb)/+layout.svelte
