+++ src/routes/(rgb)/+page.js
@@ -1,0 +1,7 @@
+
+export async function load({ fetch }) {
+	const res = await fetch('/blog/posts.json');
+	const { posts } = await res.json();
+
+	return { posts };
+}
