+++ src/routes/(rgb)/blog/+layout.svelte
