+++ src/routes/(rgb)/blog/+page.js
@@ -1,0 +1,7 @@
+
+export async function load({ fetch }) {
+	const res = await fetch('/blog/posts.json');
+	const { posts } = await res.json();
+
+	return { posts };
+}
