+++ src/routes/(rgb)/blog/+page.svelte
@@ -1,15 +1,8 @@
-<script context="module">
-	export async function load({ fetch }) {
-		const res = await fetch('/blog/posts.json');
-		const { posts } = await res.json();
 
-		return {
-			props: { posts }
-		};
-	}
-</script>
 
 <script>
+	throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
 	import BlogCard from '$lib/components/BlogCard.svelte';
 	import Breadcrumbs from '$lib/components/Breadcrumbs.svelte';
 	import SEO from '$lib/components/SEO.svelte';
