+++ src/routes/(rgb)/blog/a-cleanup-folders-script-for-your-package-json/+page.md
