+++ src/routes/(rgb)/blog/posts.json/+server.js
@@ -1,3 +1,5 @@
+throw new Error("@migration task: Update +server.js (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292701)");
+
 export const get = async () => {
 	let posts = await Promise.all(
 		Object.entries(import.meta.glob('./*.md')).map(async ([path, page]) => {
