+++ src/routes/(rgb)/privacy/+page.svelte
