+++ src/routes/(rgb)/sitemap/+page.svelte
