--- src/routes/robots.txt.js
@@ -1,13 +1,0 @@
-const BASE_URL = 'https://rgbstudios.org/';
-
-export async function get() {
-	return {
-		headers: {
-			'Cache-Control': 'max-age=0, s-maxage=3600',
-			'Content-Type': 'text/plain'
-		},
-		body: `User-agent: *
-Disallow:
-Sitemap: ${BASE_URL}sitemap.xml`.trim()
-	};
-}
